feat(favourites): Implement favourites screen
This commit introduces the Favourites screen, which displays a list of repositories that the user has marked as favourites.

Key changes include:
-   A new `FavouritesScreen` with its `ViewModel` and `State` management.
-   A `FavouriteRepositoryItem` composable to display individual favourited repositories.
-   Navigation to and from the Favourites screen and to the repository details screen.
-   Removal of the unused `SubscribedDeveloper` and `DeveloperApp` database entities.
-   Added localized strings for the "added at" timestamps in various languages.
-   Refactored the `repositoryId` in navigation and ViewModels from `Int` to `Long` for consistency.
-   Moved the `UpdatedAtFormatter` to a new `TimeFormatters.kt` file.

+++ composeApp/src/commonMain/kotlin/zed/rainxch/githubstore/feature/details/presentation/DetailsRoot.kt
@@ -66,7 +66,7 @@ import zed.rainxch.githubstore.feature.details.presentation.utils.isLiquidTopbar
 
 @Composable
 fun DetailsRoot(
-    onOpenRepositoryInApp: (repoId: Int) -> Unit,
+    onOpenRepositoryInApp: (repoId: Long) -> Unit,
     onNavigateBack: () -> Unit,
     viewModel: DetailsViewModel = koinViewModel()
 ) {
